+++ cpp/src/io/parquet/writer_impl.hpp
@@ -211,7 +211,7 @@ class writer::impl {
   Compression compression_             = Compression::UNCOMPRESSED;
   size_t max_row_group_size            = default_row_group_size_bytes;
   size_type max_row_group_rows         = default_row_group_size_rows;
-  int32_t max_page_size_bytes          = default_max_page_size_bytes;
+  size_t max_page_size_bytes           = default_max_page_size_bytes;
   size_type max_page_size_rows         = default_max_page_size_rows;
   statistics_freq stats_granularity_   = statistics_freq::STATISTICS_NONE;
   bool int96_timestamps                = false;
